# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ModelRegistry API",
    "version": "v1",
    "description": "Free unauthenticated read-only REST API for the open frontier AI model registry at https://modelregistry.tirup.in. No API keys, no signup. CORS-open (Access-Control-Allow-Origin: *). Human-readable reference: https://modelregistry.tirup.in/docs. Every record carries provenance (sources, field-level citations, verification status, changelog); benchmark figures are lab-reported unless a model page shows otherwise. Revision policy: stable kebab-case record IDs; additive fields only within v1; breaking changes ship as v2. Dataset revisions are dated (see datasetVersion); record history at https://modelregistry.tirup.in/changelog.",
    "license": {
      "name": "MIT (Open Data)",
      "url": "https://github.com/TirupMehta/ModelRegistry/blob/main/LICENSE"
    },
    "contact": {
      "name": "Tirup Mehta",
      "url": "https://tirup.in"
    }
  },
  "servers": [
    {
      "url": "https://modelregistry.tirup.in",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/api/v1/models/{id}": {
      "get": {
        "summary": "Get one tracked model by id",
        "description": "Single-record fetch so consumers can poll one model without pulling the full registry (~168KB). IDs are stable kebab-case and matched case-insensitively. Unknown ids return a JSON 404 with a lab flagship suggestion when the slug unambiguously belongs to one lab. Supports conditional requests via If-None-Match.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Stable kebab-case record id (e.g. gpt-6-astra)."
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "schema": { "type": "string" },
            "description": "ETag from a previous response. Returns 304 with an empty body when the dataset revision is unchanged."
          }
        ],
        "responses": {
          "200": {
            "description": "One registry record.",
            "headers": {
              "ETag": { "description": "Dataset revision + model id hash. Echo as If-None-Match.", "schema": { "type": "string" } },
              "Last-Modified": { "description": "Dataset revision date, not request time.", "schema": { "type": "string" } }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "model", "datasetVersion"],
                  "properties": {
                    "status": { "type": "string", "example": "success" },
                    "updatedAt": { "type": "string", "format": "date", "description": "Last real data revision (YYYY-MM-DD), never the request time.", "example": "2026-10-07" },
                    "datasetVersion": { "type": "string", "example": "2026.10.07" },
                    "methodology": { "type": "string", "example": "https://modelregistry.tirup.in/methodology" },
                    "page": { "type": "string", "example": "/models/gpt-6-astra" },
                    "url": { "type": "string", "example": "https://modelregistry.tirup.in/models/gpt-6-astra" },
                    "model": { "$ref": "#/components/schemas/ModelItem" }
                  }
                }
              }
            }
          },
          "304": {
            "description": "Not modified - dataset revision unchanged since the ETag was issued."
          },
          "404": {
            "description": "Unknown model id. Includes a flagship suggestion when the slug unambiguously belongs to one lab.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "error", "id"],
                  "properties": {
                    "status": { "type": "string", "example": "error" },
                    "error": { "type": "string", "example": "Model 'gpt-5-6-sol' not found." },
                    "id": { "type": "string" },
                    "suggestion": { "type": "string", "description": "Suggested current record id, when resolvable.", "example": "gpt-6-1-sol" },
                    "suggestionPage": { "type": "string", "example": "/models/gpt-6-1-sol" },
                    "suggestionUrl": { "type": "string", "example": "https://modelregistry.tirup.in/models/gpt-6-1-sol" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/changes": {
      "get": {
        "summary": "List registry changelog entries (incremental sync)",
        "description": "Changelog history across all records, newest first (same source as /changelog). Poll with ?since=<YYYY-MM-DD> (inclusive) to fetch only what changed, then fetch changed records via GET /api/v1/models/{id}. Unknown model/company filters return 200 with an empty list; malformed dates or limits return 400. Supports conditional requests via If-None-Match.",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "schema": { "type": "string", "format": "date" },
            "description": "Inclusive lower bound, YYYY-MM-DD (e.g. 2026-10-01)."
          },
          {
            "name": "until",
            "in": "query",
            "schema": { "type": "string", "format": "date" },
            "description": "Inclusive upper bound, YYYY-MM-DD."
          },
          {
            "name": "model",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filter to one record id (case-insensitive)."
          },
          {
            "name": "company",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filter to one laboratory id (e.g. openai)."
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 100 },
            "description": "Max entries returned. Defaults to 100, capped at 500."
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "schema": { "type": "string" },
            "description": "ETag from a previous response. Returns 304 with an empty body when nothing changed."
          }
        ],
        "responses": {
          "200": {
            "description": "Changelog entries, newest first. total is the full match count; changes is capped by limit.",
            "headers": {
              "ETag": { "description": "Revision + query hash. Echo as If-None-Match.", "schema": { "type": "string" } },
              "Last-Modified": { "description": "Dataset revision date, not request time.", "schema": { "type": "string" } }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "total", "returned", "changes", "datasetVersion"],
                  "properties": {
                    "status": { "type": "string", "example": "success" },
                    "total": { "type": "integer", "description": "Full match count before the limit cap." },
                    "returned": { "type": "integer", "description": "Entries in this response (<= limit)." },
                    "limit": { "type": "integer", "example": 100 },
                    "since": { "type": ["string", "null"], "format": "date" },
                    "until": { "type": ["string", "null"], "format": "date" },
                    "model": { "type": ["string", "null"] },
                    "company": { "type": ["string", "null"] },
                    "updatedAt": { "type": "string", "format": "date", "example": "2026-10-07" },
                    "datasetVersion": { "type": "string", "example": "2026.10.07" },
                    "methodology": { "type": "string", "example": "https://modelregistry.tirup.in/methodology" },
                    "changelog": { "type": "string", "example": "https://modelregistry.tirup.in/changelog" },
                    "changes": { "type": "array", "items": { "$ref": "#/components/schemas/ChangeEntry" } }
                  }
                }
              }
            }
          },
          "304": {
            "description": "Not modified - dataset revision and query unchanged since the ETag was issued."
          },
          "400": {
            "description": "Malformed since/until date or limit.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "error"],
                  "properties": {
                    "status": { "type": "string", "example": "error" },
                    "error": { "type": "string", "example": "Invalid 'since' date 'last-week'. Expected YYYY-MM-DD." }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "SourceRef": {
        "type": "object",
        "description": "One cited primary source behind a record.",
        "required": ["url", "publisher", "title", "accessedAt", "sourceType"],
        "properties": {
          "url": { "type": "string", "format": "uri", "example": "https://mistral.ai/news/mistral-large-4/" },
          "publisher": { "type": "string", "example": "Mistral AI" },
          "title": { "type": "string", "example": "Mistral Large 4 announcement" },
          "publishedAt": { "type": "string", "format": "date", "description": "Source publication date, when directly observed." },
          "accessedAt": { "type": "string", "format": "date", "description": "Date the registry last confirmed the source.", "example": "2026-10-07" },
          "sourceType": { "type": "string", "enum": ["announcement", "api-docs", "pricing", "model-card", "paper", "weights", "benchmark", "console"] },
          "live": { "type": "boolean", "description": "True when the source returned HTTP 2xx to the automated check." },
          "note": { "type": "string", "description": "Access limits or source weaknesses, e.g. bot-gated pages." }
        }
      },
      "ModelItem": {
        "type": "object",
        "description": "One registry record. Benchmark figures are lab-reported, not independently reproduced.",
        "required": ["id", "companyId", "name", "releaseDate", "pricing", "sources", "fieldSources", "lastVerifiedAt", "verificationStatus", "changeLog"],
        "properties": {
          "id": { "type": "string", "description": "Stable kebab-case record ID.", "example": "mistral-large-4" },
          "companyId": { "type": "string", "example": "mistral" },
          "companyName": { "type": "string", "example": "Mistral AI" },
          "name": { "type": "string", "example": "Mistral Large 4" },
          "releaseDate": { "type": "string", "format": "date" },
          "isCompanyFlagship": { "type": "boolean", "description": "Lab's primary general-purpose model. Requires fresh, live evidence." },
          "isLatestCheckpoint": { "type": "boolean", "description": "Lab's newest shipped release. Requires fresh, live evidence." },
          "contextWindowTokens": { "type": "integer", "description": "0 means undisclosed or non-token billing." },
          "pricing": {
            "type": "object",
            "required": ["input", "output"],
            "properties": {
              "input": { "type": "number", "description": "USD per billing unit." },
              "output": { "type": "number" }
            }
          },
          "pricingUnit": { "type": "string", "description": "Only 'per second'; absent means per-1M-tokens." },
          "sources": { "type": "array", "items": { "$ref": "#/components/schemas/SourceRef" } },
          "fieldSources": {
            "type": "object",
            "description": "Per-field citations keyed by releaseDate, contextWindow, maxOutputTokens, parameters, license, pricing, modalities, benchmarks, isCompanyFlagship, isLatestCheckpoint. Missing keys mean no citable source yet.",
            "additionalProperties": { "type": "array", "items": { "$ref": "#/components/schemas/SourceRef" } }
          },
          "lastVerifiedAt": { "type": "string", "format": "date" },
          "verificationStatus": { "type": "string", "enum": ["verified", "partially_verified", "unverified", "retired"] },
          "changeLog": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["date", "summary"],
              "properties": {
                "date": { "type": "string", "format": "date" },
                "summary": { "type": "string" },
                "sources": { "type": "array", "items": { "type": "string", "format": "uri" } }
              }
            }
          },
          "supersededBy": { "type": "string", "description": "Replacement record id, when retired." }
        }
      },
      "ChangeEntry": {
        "type": "object",
        "description": "One dated changelog entry from a model's changeLog history, with its owning record attached.",
        "required": ["date", "summary", "modelId", "modelName", "companyId"],
        "properties": {
          "date": { "type": "string", "format": "date", "example": "2026-10-07" },
          "summary": { "type": "string", "example": "Audit correction: re-sourced announcement to canonical blog URL." },
          "sources": { "type": "array", "items": { "type": "string", "format": "uri" } },
          "modelId": { "type": "string", "example": "gemini-3-8-flash" },
          "modelName": { "type": "string", "example": "Gemini 3.8 Flash" },
          "companyId": { "type": "string", "example": "google" },
          "companyName": { "type": "string", "example": "Google DeepMind" },
          "page": { "type": "string", "example": "/models/gemini-3-8-flash" },
          "url": { "type": "string", "example": "https://modelregistry.tirup.in/models/gemini-3-8-flash" }
        }
      }
    }
  }
}